+++ src/main/java/com/uranus/taskmanager/api/security/session/SessionManager.java
@@ -1,6 +1,4 @@
-package com.uranus.taskmanager.api.security.authentication.session;
-
-import static com.uranus.taskmanager.api.security.authentication.session.SessionAttributes.*;
+package com.uranus.taskmanager.api.security.session;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
@@ -28,9 +26,9 @@ public class SessionManager {
 
 	// 로그인 세션 설정, 관리
 	public void createLoginSession(HttpSession session, LoginResponse loginResponse) {
-		session.setAttribute(LOGIN_MEMBER_ID, loginResponse.getId());
-		session.setAttribute(LOGIN_MEMBER_LOGIN_ID, loginResponse.getLoginId());
-		session.setAttribute(LOGIN_MEMBER_EMAIL, loginResponse.getEmail());
+		session.setAttribute(SessionAttributes.LOGIN_MEMBER_ID, loginResponse.getId());
+		session.setAttribute(SessionAttributes.LOGIN_MEMBER_LOGIN_ID, loginResponse.getLoginId());
+		session.setAttribute(SessionAttributes.LOGIN_MEMBER_EMAIL, loginResponse.getEmail());
 		log.info("Login session created for member ID: {}", loginResponse.getId());
 	}
 
@@ -43,20 +41,20 @@ public class SessionManager {
 
 	public Optional<Long> getLoginMemberId(HttpSession session) {
 		return Optional.ofNullable(session)
-			.map(s -> (Long)s.getAttribute(LOGIN_MEMBER_ID));
+			.map(s -> (Long)s.getAttribute(SessionAttributes.LOGIN_MEMBER_ID));
 	}
 
 	// 업데이트 권한 설정, 관리
 	public void createUpdatePermission(HttpSession session) {
 		LocalDateTime expiresAt = LocalDateTime.now().plusMinutes(UPDATE_PERMISSION_MINUTES);
-		session.setAttribute(UPDATE_AUTH, true);
-		session.setAttribute(UPDATE_AUTH_EXPIRES_AT, expiresAt);
+		session.setAttribute(SessionAttributes.UPDATE_AUTH, true);
+		session.setAttribute(SessionAttributes.UPDATE_AUTH_EXPIRES_AT, expiresAt);
 		log.info("Update permission created, expires at: {}", expiresAt);
 	}
 
 	// 세션 정보 업데이트
 	public void updateSessionEmail(HttpSession session, String newEmail) {
-		session.setAttribute(LOGIN_MEMBER_EMAIL, newEmail);
+		session.setAttribute(SessionAttributes.LOGIN_MEMBER_EMAIL, newEmail);
 		log.info("Session email updated to: {}", newEmail);
 	}
