+++ src/test/java/com/uranus/taskmanager/helper/ControllerTestHelper.java
@@ -21,8 +21,8 @@ import com.uranus.taskmanager.api.member.service.MemberQueryService;
 import com.uranus.taskmanager.api.member.service.MemberService;
 import com.uranus.taskmanager.api.security.authentication.presentation.controller.AuthenticationController;
 import com.uranus.taskmanager.api.security.authentication.service.AuthenticationService;
-import com.uranus.taskmanager.api.security.authentication.session.SessionManager;
-import com.uranus.taskmanager.api.security.authentication.session.SessionValidator;
+import com.uranus.taskmanager.api.security.session.SessionManager;
+import com.uranus.taskmanager.api.security.session.SessionValidator;
 import com.uranus.taskmanager.api.workspace.domain.repository.WorkspaceRepository;
 import com.uranus.taskmanager.api.workspace.presentation.controller.WorkspaceController;
 import com.uranus.taskmanager.api.workspace.service.WorkspaceCommandService;
